apps/triage/requirements.txt
pandas==3.0.2
psycopg2-binary==2.9.11
PyYAML==6.0.3
